+++ components/ui/atoms/index.ts
@@ -2,3 +2,4 @@ export { default as Button } from './Button'
 export { default as StyledText } from './Text'
 export { default as View } from './View'
 export { default as Checkbox } from './Checkbox'
+export { default as Textfield } from './Textfield'
